Simplify mocking ledger storage using mockito
There are a lot of dummy implementation of ledger storage with most of methods unimplemented. When there are singatures changed
in ledger storage interface, it requires to update many files. Change to use mockito to provide the mocked implementation with
only implementing needed methods. It makes code cleaner and easier to maintain.

One small change to make GabageCollector to be functional interfaces.